attributes: using intrinsic predefined methods
With this change, attribute getters and setters are defined
as intrinsic methods of the class, that can in principle be exposed
or not (using a separate `impl` block) to Python.

At this stage, it is necessary not to expose the setter,
because of its return type and the fact that `()` doesn't implement
`ToPythonObject`.

It may feel more rustic to change the incoming type declaration from
`setter(TYPE)` to `setter<TYPE>` but I'm worried it might mislead people
into thinking that several types could be allowed.